fix: git-ignore static asset artifacts whether they are links or dirs
Remove the trailing slashes from the .gitignore entries for static asset
build artifacts. With these slashes, only directory contents are
ignored. Without these slashes, the artifact is ignored whether it is a
directory *or* an actual file (particularly, in our case: a symlink).

This allows us to build edx-platform static assets into a separate
directory, and then symlink to them from edx-platform.