refactor: extract save_xblock_to_user_clipboard Python API
Previously, the logic for copying an XBlock's OLX to the clipboard
was directly in a POST view handler. This commit extracts it into a
Python API function.

In order to keep models contained within api.py, we move some logic from
the UserClipboard model to the UserClipboardData object and change the
UserClipboardSerializer to accept the latter instead of the former. As
an unavoidable side effect, this means that the source_context_title of
the clipboard will not be available in Django admin.

In addition to following architectural guidelines [1], this change will
help me develop the Content Libraries Relaunch, as I need a way to
locally test library content reference via copy-paste while the UX for
it is still being developed.

[1] https://docs.openedx.org/projects/openedx-proposals/en/latest/best-practices/oep-0049-django-app-patterns.html